fix: split prompts separately
